Improved UserProfile Edit & Language Features
